cloud-functions/textspeech/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-texttospeech